Replace reqwest HTTP client with ureq
This is a working implementation of this now. The reqwest library made it self seem simple and easy to use, but it was a massive pain in the ass. ureq actually is painless to use, although the downside is that it doesn't use async Rust, but rather it is simply blocking. Although this may make this app run slightly slower it shouldn't be a huge issue given the scale of the expected input.